# a7_openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: 'LBAW2144 - Sustainlbaw'
  description: 'Web Resources Specification (A7)'
servers:
# Added by API Auto Mocking Plugin
- description: SwaggerHub API Auto Mocking
  url: https://virtserver.swaggerhub.com/tiago-falves/SustainlbawLbaw44/1.0.0
- url: http://lbaw2144.lbaw-prod.fe.up.pt
  description: Production server
- url: http://localhost:8000
  description: Development server
...
paths:
...
  /login:
    get:
      operationId: M0101
      summary: Showing login form
      description: 'Login form. Acess: PUB'
      tags:
        - 'M01: Authentication & User'
      responses:
        '200':
          description: 'OK. Show [UI06](https://git.fe.up.pt/lbaw/lbaw2021/lbaw2144/-/wikis/er#ui06-sign-in)'
        '401':
          description: 'ERROR. Try again'
        '404':
          description: Login not Found
        '429':
          description: Too Many Requests
        default:
          description: Default
          
    post:
      operationId: M0102
      summary: 'Login Action'
      description: 'Processes the login form submission. Access: PUB'
      tags:
        - 'M01: Authentication & User'
      #TODO Isto nao devia estar nos parameters?
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:          # <!--- form field name
                  type: string
                password:    # <!--- form field name
                  type: string
              required:
                - email
                - password
 
      responses:
        '302':
          description: 'Redirect after processing the login credentials.'
          headers:
            Location:
              schema:
                type: string
              examples:
                '302 Success':
                    description: 'Successful authentication. Redirect to user profile.'
                    value: '/'
                '302 Error':
                  description: 'Failed authentication. Redirect to login form.'
                  value: '/login'
        '204':
          description: Created
        '429':
          description: Too Many Requests
        default:
          description: Default
...
  /blog:
    get:
      operationId: M0301
      summary: All blog posts
      description: 'Getting all blog posts. Acess: PUB'
      tags:
        - 'M03: Blog'
      responses:
        '200':
          description: 'OK. All blog posts returned'
        '401':
          description: 'ERROR. Try again'
        '404':
          description: Not Found
        '429':
          description: Too Many Requests
        default:
          description: Default
    
    post:
      operationId: M0302
      summary: Add blog post
      description: 'Adding a new blog post. Acess: EDI'
      tags:
        - 'M03: Blog'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                title:          # <!--- form field name
                  type: string
                content:
                  type: string
                image_path:
                  type: string
                author:
                  type: string
                editor:
                  type: integer
              required:
                - title
                - content
                - image_path
                - author
                - editor
      responses:
        '201':
          description: 'Post created sucessfully'
        '429':
          description: Too Many Requests
        default:
          description: Default
...
  /blog/{id}/edit:    
    get:
      operationId: M0308
      summary: Find blog post edit form by id
      description: 'Getting the edit blog post page. Acess: PUB'
      tags:
        - 'M03: Blog'
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: 'OK. Edit Blog post page returned'
        '401':
          description: 'ERROR. Try again'
        '404':
          description: Not Found
        '429':
          description: Too Many Requests
        default:
          description: Default
...